requirements.txt
appnope==0.1.0
attrs==19.3.0
backcall==0.2.0
beautifulsoup4==4.9.1
billiard==3.6.3.0
bleach==3.1.5
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
coala-utils==0.5.1
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
Flask==1.1.2
googletrans==3.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.6.30
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
importlib-metadata==1.7.0
ipykernel==5.3.0
ipython==7.16.1
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.17.1
Jinja2==2.11.2
joblib==0.16.0
jsonschema==3.2.0
jupyter-client==6.1.5
jupyter-core==4.6.3
lxml==4.5.1
MarkupSafe==1.1.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.7
newsapi-python==0.2.6
nltk==3.5
notebook==6.0.3
numpy==1.19.0
oauthlib==3.1.0
packaging==20.4
pandas==1.0.5
pandocfilters==1.4.2
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
prometheus-client==0.8.0
prompt-toolkit==3.0.5
ptyprocess==0.6.0
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.1
pyzmq==19.0.1
regex==2020.6.8
requests==2.24.0
requests-oauthlib==1.3.0
rfc3986==1.4.0
scikit-learn==0.23.1
scipy==1.5.0
Send2Trash==1.5.0
six==1.15.0
sklearn==0.0
sniffio==1.1.0
soupsieve==2.0.1
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.1.0
tornado==6.0.4
tqdm==4.47.0
traitlets==4.3.3
tweepy==3.8.0
twitterscraper==1.4.0
urllib3==1.25.9
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
zipp==3.1.0
